Move RRN functions to internal gke package
The Relative Resource Name helper functions are closely related to using
the GKE API and not generic utilities, so move them into the internal
gke package.
