+++ internal/gke/relative_resource_name.go
@@ -1,4 +1,4 @@
-package utils
+package gke
 
 import (
 	"fmt"
