+++ internal/gke/update.go
@@ -56,7 +56,7 @@ func UpdateMasterKubernetesVersion(ctx context.Context, client *gkeapi.Service, 
 			Locations.
 			Clusters.
 			Update(
-				utils.ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
+				ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
 				&gkeapi.UpdateClusterRequest{
 					Update: &gkeapi.ClusterUpdate{
 						DesiredMasterVersion: *config.Spec.KubernetesVersion,
@@ -123,7 +123,7 @@ func UpdateClusterAddons(ctx context.Context, client *gkeapi.Service, config *gk
 			Locations.
 			Clusters.
 			Update(
-				utils.ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
+				ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
 				&gkeapi.UpdateClusterRequest{
 					Update: clusterUpdate,
 				},
@@ -192,7 +192,7 @@ func UpdateMasterAuthorizedNetworks(
 			Locations.
 			Clusters.
 			Update(
-				utils.ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
+				ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
 				&gkeapi.UpdateClusterRequest{
 					Update: clusterUpdate,
 				},
@@ -241,7 +241,7 @@ func UpdateLoggingMonitoringService(
 			Locations.
 			Clusters.
 			Update(
-				utils.ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
+				ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
 				&gkeapi.UpdateClusterRequest{
 					Update: clusterUpdate,
 				},
@@ -271,7 +271,7 @@ func UpdateNetworkPolicyEnabled(
 			Locations.
 			Clusters.
 			SetNetworkPolicy(
-				utils.ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
+				ClusterRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName),
 				&gkeapi.SetNetworkPolicyRequest{
 					NetworkPolicy: &gkeapi.NetworkPolicy{
 						Enabled:  *config.Spec.NetworkPolicyEnabled,
@@ -319,7 +319,7 @@ func UpdateNodePoolKubernetesVersionOrImageType(
 			Clusters.
 			NodePools.
 			Update(
-				utils.NodePoolRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName, *nodePool.Name),
+				NodePoolRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName, *nodePool.Name),
 				updateRequest,
 			).Context(ctx).
 			Do()
@@ -352,7 +352,7 @@ func UpdateNodePoolSize(
 		Clusters.
 		NodePools.
 		SetSize(
-			utils.NodePoolRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName, *nodePool.Name),
+			NodePoolRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName, *nodePool.Name),
 			&gkeapi.SetNodePoolSizeRequest{
 				NodeCount: *nodePool.InitialNodeCount,
 			},
@@ -400,7 +400,7 @@ func UpdateNodePoolAutoscaling(
 			Clusters.
 			NodePools.
 			SetAutoscaling(
-				utils.NodePoolRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName, *nodePool.Name),
+				NodePoolRRN(config.Spec.ProjectID, config.Spec.Region, config.Spec.ClusterName, *nodePool.Name),
 				updateRequest,
 			).Context(ctx).
 			Do()
